requirements.txt
black==24.10.0
flake8==7.1.1
isort==5.13.2
pre-commit==4.0.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-telegram-bot
python-decouple==3.8
redis==5.2.1
requests==2.32.3
khayyam
